Add support for multiple query definitions
Allow multiple queries to be defined in the bruv configuration files
(config option 'queries'). Allow a query to be specified via command
line.

Queries specified via the command line are first searched through the
defined queries, and if not found, used as is. If no query is specified
from the command line, the 'query' config parameter is used as default.
It is also searched first in the 'queries' dictionary, to allow aliasing.
This exists to allow backwards compatibility.